Rename interface_registry Trim down interface_registry
