--- test/testdata/codec.go
@@ -1,28 +1,0 @@
-package testdata
-
-import (
-	types "github.com/cosmos/gogoproto/types/any"
-)
-
-func NewTestInterfaceRegistry() types.InterfaceRegistry {
-	registry := types.NewInterfaceRegistry()
-	RegisterInterfaces(registry)
-	return registry
-}
-
-func RegisterInterfaces(registry types.InterfaceRegistry) {
-	registry.RegisterInterface("Animal", (*Animal)(nil))
-	registry.RegisterImplementations(
-		(*Animal)(nil),
-		&Dog{},
-		&Cat{},
-	)
-	registry.RegisterImplementations(
-		(*HasAnimalI)(nil),
-		&HasAnimal{},
-	)
-	registry.RegisterImplementations(
-		(*HasHasAnimalI)(nil),
-		&HasHasAnimal{},
-	)
-}
